@@ -5,9 +5,9 @@ namespace App\Http\Controllers;
 use App\Models\Food;
 use App\Models\Recipe;
 use App\Search\Ingredient;
-use ElasticScoutDriverPlus\Builders\MultiMatchQueryBuilder;
-use ElasticScoutDriverPlus\Builders\TermsQueryBuilder;
-use ElasticScoutDriverPlus\Support\Query;
+use Elastic\ScoutDriverPlus\Builders\MultiMatchQueryBuilder;
+use Elastic\ScoutDriverPlus\Builders\TermsQueryBuilder;
+use Elastic\ScoutDriverPlus\Support\Query;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
