@@ -9,7 +9,7 @@ use App\Models\Traits\Sluggable;
 use App\Models\Traits\Taggable;
 use App\Support\Number;
 use App\Support\Nutrients;
-use ElasticScoutDriverPlus\Searchable;
+use Elastic\ScoutDriverPlus\Searchable;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasMany;
